+++ e2e/packages/sync-test/data/setContractData.ts
@@ -3,7 +3,7 @@ import { Data } from "./types";
 import { encodeTestData } from "./encodeTestData";
 import { callWorld } from "./callWorld";
 import { stringToBytes16 } from "@latticexyz/utils";
-import { toHex } from "viem";
+import { Hex, toHex } from "viem";
 
 /**
  * Writes contract data by calling `world.setRecord` via the client
@@ -13,12 +13,12 @@ export async function setContractData(page: Page, data: Data) {
   const promises: Promise<unknown>[] = [];
   for (const [table, records] of Object.entries(encodedData)) {
     for (const record of records) {
-      const promise = callWorld(page, "setRecord(bytes16,bytes16,bytes32[],bytes)", [
+      const promise = await callWorld(page, "setRecord", [
         // TODO: add support for multiple namespaces after https://github.com/latticexyz/mud/issues/994 is resolved
         toHex(stringToBytes16("")),
         toHex(stringToBytes16(table)),
-        record.key,
-        record.value,
+        record.key as Hex[],
+        record.value as Hex,
       ]);
 
       // Wait for transactions to be confirmed